Added initial framework/protocol data validation and cleansing support.
Before requesting the LDP4j application to modify a resource, the LDP4j
application kernel verifies that the client has not messed with the
server managed properties, then removes all those properties from the
input data (as they are not to be consumed by the application) and
then handles the course of action to the application.

Extended Property API to support checking whether a property contains a
literal or an identified individual.

Added new utility methods for dealing with data sets, including a simple
data validation framework.