Started new API for handling data transformations.
The purpose of the new API is to be able to separate entities from
identities, so that the resolution process can be made much more
flexible, allowing to deal with corner cases like handling external
individuals which should be published in a URI relative to a managed
individual.

In addition, the new API will make a lot of emphasis on usability,
making it clear and easy to use altogether.